server/requirements.txt
bert_extractive_summarizer==0.10.1
fitz==0.0.1.dev2
Flask==2.2.2
summarizer==0.0.7
PyMuPDF==1.21.0
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
scapy==3.4.3
spacy-legacy=3.0.10
spacy-loggers=1.0.4